Refactor storage, environment, CLI
This is a batch of changes with the goal of simplifying recipe writing.

* Handle flow environment and storage outside of the recipe files.
* Moved to `GitHub` storage. Flows are now stored as files and loaded
  into the environment.
* Moved `Environment` to the base pipeline

See https://github.com/TomAugspurger/example-pipeline/blob/main/recipe/pipeline.py
for an example pipeline run with this setup. The primary change is the
*requirement* that the top-level of the module contains a `Flow`
instance named `flow`.

Additionally, the `pangeo-forge` CLI is updated to enable running flows
with the new structure. The CLI handles

1. Linting recipe pipelines.
2. Attaching Prefect environments and Storage to `Flow`s.
3. Registering flows with prefect.