Prevent operator fusion for every stage of the Beam executor.
In an end-to-end experiment, I noticed that stages were still being fused. This caused errors to occur, namely that the Zarr metadata file was not created before one of the later steps (I think `store_chunk`) that needed it.

By adding a `beam.Reshuffle()` step to the end of every stage, we ensure that fusion doesn't occur and that stages do execute serially.