Make FSSpecTarget be 'pathlike'
The goal here is for recipe authors to be able to
pass *strings* to `target` parameters of StoreToZarr to specify
where the data should be for this specific recipe, and we can
inject the prefix with pangeo-forge-runner.

The primary use case is when we have a dictionary of recipes
being returned from a single recipe file. The *name* of the recipe
is unfortunately not really available during AST parsing, so the
suffix the data should be put in needs to be explicitly specified.