Add target_subpath to StoreToZarr
When generating multiple Zarr datasets from the same recipe,
we want to support specifying *where* each dataset will go inside
the same target, as target is set by pangeo-forge-runner. Without,
this subpath is specified as the *name* of the recipe itself. However,
due to how ast rewriting works, we don't actually have the name of
the recipe when constructing our beam pipeline, so we can not use
that. So instead, it would have to be explicitly specified as a
subpath inside the target.